requirements/base.txt
Django==1.6.2
bpython==0.12
django-braces==1.2.2
django-model-utils==1.5.0
logutils==0.3.3
South==0.8.2

django-treemenus==0.9.0
django-photologue==2.7
django-tagging==0.3.2
django-tinymce==1.5.2
psycopg2==2.5.2
